+++ changelog.md
@@ -8,10 +8,13 @@
 
 ### Changed
 
+- Switch the DataView to script-based config for big configurations
 - Hide hidden sections in DataView
 - Optimized the usage of MultiField
 - Use identical styling for legends on each field in DataView
 - Extended the support for WP_List_Table in DataView
+- Optimized error handling for faulty JSON configurations
+- Optimized styling of tables in classic view
 
 ## [3.1.2] - 15.08.2026